@@ -28,7 +28,7 @@ bl_info = {
     "name": "Ekur",
     "description": "A multi-purpose importer for Halo Infinite.",
     "author": "Surasia",
-    "version": (0, 4, 0),
+    "version": (0, 5, 0),
     "blender": (4, 3, 0),
     "category": "Import-Export",
     "support": "COMMUNITY",
