@@ -1,7 +1,6 @@
 /* SPDX-License-Identifier: GPL-3.0-or-later */
 /* Copyright © 2025 Surasia */
 use infinite_rs::tag::types::common_types::FieldRealQuaternion;
-use nalgebra::{Matrix4, Vector4};
 
 use crate::definitions::render_model::NodeBlock;
 
@@ -9,11 +8,12 @@ use crate::definitions::render_model::NodeBlock;
 pub(super) struct Bone {
     pub(super) name: i32,
     pub(super) parent_index: i32,
-    pub(super) local_transform: [[f32; 4]; 4],
+    pub(super) rotation_matrix: [[f32; 4]; 4],
+    pub(super) transformation_matrix: [[f32; 4]; 4],
     pub(super) world_transform: [[f32; 4]; 4],
 }
 
-pub fn create_matrix_from_quaternion(quaternion: &FieldRealQuaternion) -> Matrix4<f32> {
+pub(super) fn create_matrix_from_quaternion(quaternion: &FieldRealQuaternion) -> [[f32; 4]; 4] {
     let xx = quaternion.x * quaternion.x;
     let yy = quaternion.y * quaternion.y;
     let zz = quaternion.z * quaternion.z;
@@ -25,12 +25,12 @@ pub fn create_matrix_from_quaternion(quaternion: &FieldRealQuaternion) -> Matrix
     let yz = quaternion.y * quaternion.z;
     let wx = quaternion.x * quaternion.w;
 
-    Matrix4::from_columns(&[
-        Vector4::new(1.0 - 2.0 * (yy + zz), 2.0 * (xy + wz), 2.0 * (xz - wy), 0.0),
-        Vector4::new(2.0 * (xy - wz), 1.0 - 2.0 * (zz + xx), 2.0 * (yz + wx), 0.0),
-        Vector4::new(2.0 * (xz + wy), 2.0 * (yz - wx), 1.0 - 2.0 * (yy + xx), 0.0),
-        Vector4::new(0.0, 0.0, 0.0, 1.0),
-    ])
+    [
+        [1.0 - 2.0 * (yy + zz), 2.0 * (xy + wz), 2.0 * (xz - wy), 0.0],
+        [2.0 * (xy - wz), 1.0 - 2.0 * (zz + xx), 2.0 * (yz + wx), 0.0],
+        [2.0 * (xz + wy), 2.0 * (yz - wx), 1.0 - 2.0 * (yy + xx), 0.0],
+        [0.0, 0.0, 0.0, 1.0],
+    ]
 }
 
 pub(super) fn create_node(node: &NodeBlock) -> Bone {
@@ -47,25 +47,19 @@ pub(super) fn create_node(node: &NodeBlock) -> Bone {
     let up = &node.inverse_up;
     let po = &node.inverse_position;
 
-    let rot_matrix = create_matrix_from_quaternion(rot2);
-    let trans_matrix = Matrix4::from_columns(&[
-        Vector4::new(1.0, 0.0, 0.0, 0.0),
-        Vector4::new(0.0, 1.0, 0.0, 0.0),
-        Vector4::new(0.0, 0.0, 1.0, 0.0),
-        Vector4::new(pos2.x, pos2.y, pos2.z, 1.0),
-    ]);
+    bone.rotation_matrix = create_matrix_from_quaternion(rot2);
+    bone.transformation_matrix = [
+        [1.0, 0.0, 0.0, 0.0],
+        [0.0, 1.0, 0.0, 0.0],
+        [0.0, 0.0, 1.0, 0.0],
+        [pos2.x, pos2.y, pos2.z, 1.0],
+    ];
 
-    let local_transform = trans_matrix * rot_matrix;
-
-    let world_transform = Matrix4::from_columns(&[
-        Vector4::new(fw.x, fw.y, fw.z, 0.0),
-        Vector4::new(lf.x, lf.y, lf.z, 0.0),
-        Vector4::new(up.x, up.y, up.z, 0.0),
-        Vector4::new(po.x, po.y, po.z, 1.0),
-    ]);
-
-    let world_transform = world_transform.try_inverse().unwrap_or(Matrix4::identity());
-    bone.local_transform = local_transform.data.0;
-    bone.world_transform = world_transform.data.0;
+    bone.world_transform = [
+        [fw.x, fw.y, fw.z, 0.0],
+        [lf.x, lf.y, lf.z, 0.0],
+        [up.x, up.y, up.z, 0.0],
+        [po.x, po.y, po.z, 1.0],
+    ];
     bone
 }
